exec_lib.sh: retesting failures for debugging
To debug a failure one usually needs to execute the acvpaparser
command that failed (the exact command is already part of the
output). However, if there are many failures the process of
getting all results needed might be tedious.

This change adds support for automated re-execution of all
failures with given verbosity level (DEBUG_FAILED_VERBOSITY)
and storing the output in a given file (DEBUG_FAILED_OUTPUT).

If DEBUG_FAILED_OUTPUT is not set, no re-execution is done.
If DEBUG_FAILED_VERBOSITY is not set, the verbosity level is -v.

Signed-off-by: Stephan Mueller <smueller@chronox.de>