add support for AES-XTS data unit length
Add support for AES-XTS algorithm with data unit length different than payload
length. For ACVP-AES-XTS revision 2.0, which deprecated the previous one,
registering the algorithm with the field "dataUnitLenMatchesPayload" is
required. This change allows successful execution of the tests registered with
the above field set to false, and it is preserving compatibility when the value
is true.

Signed-off-by: Karol Brzuskiewicz <kabr@arista.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>