Casting / Conversion Bug on Some Platforms in the SHA Parser
This isn't an issue on 64-bit targets, but for 32-bit targets this could
cause a loss of precision. This was caught by our compiler due to
"-Wshorten-64-to-32" being enabled.

Based on reading the spec for the long data tests, it seems like the
ldt_expansion_size value could be as high as 2^36. If that got cast to a
size_t (even with the division by 8), that would become 0.

Reported-by: Mike P
Signed-off-by: Stephan Mueller <stephan.mueller@atsec.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>