backend_openssl3 support to use default provider
This adds support for the openssl3 backend to use the default provider.

Why is this needed?
The openssl3 backend normally loads the fips provider module.  However,
using fips is not mandatory and the tool can be used with the default
algorithm provider.  In this case, it is inappropriate to load the fips
module (if it even exists on the system) and to enable fips mode.

Building the tool with the new preprocessor symbol
OPENSSL3_BACKEND_USE_DEFAULT_PROVIDER defined avoids using fips mode.
To use this, build with CFLAGS containing
"-DOPENSSL3_BACKEND_USE_DEFAULT_PROVIDER".
The default is to use fips mode.

Tested:
Builds when the new symbol is not defined.
Builds with the new symbol defined, does not load the fips module, and
succesfully runs the tests.

Signed-off-by: Joseph Reynolds <joseph.reynolds1@ibm.com>